third_party/blink/web_tests/external/wpt/webrtc/RTCDataChannel-binaryType.window-expected.txt

This is a testharness.js-based test.
[FAIL] Default binaryType value
  assert_equals: dc.binaryType should be 'blob' expected "blob" but got "arraybuffer"
[FAIL] Setting binaryType to 'blob' should succeed
  Failed to set the 'binaryType' property on 'RTCDataChannel': Blob support not implemented yet
Harness: the test ran to completion.
